refactor(hooks/use-handle-sign-in): move sign-in logic to hooks
